requirements.txt
fastapi==0.115.6
uvicorn==0.22.0
httpx==0.28.1
Pillow==11.0.0
pytesseract==0.3.13
langgraph==0.3.5
python-multipart==0.0.6
pydantic==2.10.3
python-dotenv==1.0.1
google-genai==1.5.0